+++ tests/src/Supply/Domain/ValueObject/SizeTest.php
@@ -43,4 +43,31 @@ final class SizeTest extends TestCase
         $this->assertEquals([728, 0], Size::toDimensions('728'));
         $this->assertEquals([0, 0], Size::toDimensions(''));
     }
+
+    public function testAspect(): void
+    {
+        $this->assertEquals('4:3', Size::getAspect(320, 240));
+        $this->assertEquals('6:5', Size::getAspect(300, 250));
+        $this->assertEquals('', Size::getAspect(320, 0));
+        $this->assertEquals('', Size::getAspect(0, 240));
+    }
+
+    public function testFindBestFit(): void
+    {
+        $this->assertContains('300x250', Size::findBestFit(300, 250, 1));
+        $this->assertContains('336x280', Size::findBestFit(330, 270, 1));
+    }
+
+    public function testFindMatching(): void
+    {
+        $this->assertEmpty(Size::findMatching(1, 1));
+        $this->assertEmpty(Size::findMatching(300, 0));
+        $this->assertEmpty(Size::findMatching(300, 10));
+        $this->assertContains('300x250', Size::findMatching(300, 250));
+        $this->assertContains('300x250', Size::findMatching(320, 240));
+        $this->assertContains('300x250', Size::findMatching(1920, 1080));
+        $this->assertContains('300x600', Size::findMatching(1080, 1920));
+        $this->assertContains('300x250', Size::findMatching(4000, 3000));
+        $this->assertContains('300x600', Size::findMatching(3000, 4000));
+    }
 }